[raftstore] move snapshot generation out of the region worker
Issue Number: Close #16001

This is the second attempt to fix #16001 and #12587. The previous fix
(#17121) was reverted (#17275) due to a concurrency bug between
snapshot apply and destroy.

To fix the issue where slow region destruction can block snapshot
generation, this PR moves the snapshot generation logic out of the
region worker and into a separate worker. This is a simpler approach
than the earlier attempt because it doesn't deal with interactions
between snapshot apply and destroy. Since snapshot generation has
always been an independent task handled by its own thread pool, this
change does not add significant complexity.

However, one inefficiency remains: the snap gen worker mainly forwards
tasks to the snap gen thread pool, becoming a thread that doesn't do
much work. One optimization to consider is to send tasks directly to
the thread pool without using the snap gen worker as an intermediary.
