+++ components/raftstore/src/store/fsm/apply.rs
@@ -95,7 +95,7 @@ use crate::{
             self, admin_cmd_epoch_lookup, check_flashback_state, check_req_region_epoch,
             compare_region_epoch, ChangePeerI, ConfChangeKind, KeysInfoFormatter,
         },
-        Config, RegionSnapshot, RegionTask, WriteCallback,
+        Config, RegionSnapshot, SnapGenTask, WriteCallback,
     },
     Error, Result,
 };
@@ -397,7 +397,7 @@ where
     timer: Option<Instant>,
     host: CoprocessorHost<EK>,
     importer: Arc<SstImporter<EK>>,
-    region_scheduler: Scheduler<RegionTask<EK::Snapshot>>,
+    snap_gen_scheduler: Scheduler<SnapGenTask<EK::Snapshot>>,
     router: ApplyRouter<EK>,
     notifier: Box<dyn Notifier<EK>>,
     engine: EK,
@@ -482,7 +482,7 @@ where
         tag: String,
         host: CoprocessorHost<EK>,
         importer: Arc<SstImporter<EK>>,
-        region_scheduler: Scheduler<RegionTask<EK::Snapshot>>,
+        snap_gen_scheduler: Scheduler<SnapGenTask<EK::Snapshot>>,
         engine: EK,
         router: ApplyRouter<EK>,
         notifier: Box<dyn Notifier<EK>>,
@@ -498,7 +498,7 @@ where
             timer: None,
             host,
             importer,
-            region_scheduler,
+            snap_gen_scheduler,
             engine,
             router,
             notifier,
@@ -3696,14 +3696,14 @@ impl GenSnapTask {
         kv_snap: EK::Snapshot,
         last_applied_term: u64,
         last_applied_state: RaftApplyState,
-        region_sched: &Scheduler<RegionTask<EK::Snapshot>>,
+        snap_gen_sched: &Scheduler<SnapGenTask<EK::Snapshot>>,
     ) -> Result<()>
     where
         EK: KvEngine,
     {
         self.index
             .store(last_applied_state.applied_index, Ordering::SeqCst);
-        let snapshot = RegionTask::Gen {
+        let snapshot = SnapGenTask::Gen {
             region_id: self.region_id,
             notifier: self.snap_notifier,
             for_balance: self.for_balance,
@@ -3715,7 +3715,7 @@ impl GenSnapTask {
             kv_snap,
             to_store_id: self.to_peer.store_id,
         };
-        box_try!(region_sched.schedule(snapshot));
+        box_try!(snap_gen_sched.schedule(snapshot));
         Ok(())
     }
 }
@@ -4262,7 +4262,7 @@ where
             apply_ctx.engine.snapshot(None),
             self.delegate.applied_term,
             self.delegate.apply_state.clone(),
-            &apply_ctx.region_scheduler,
+            &apply_ctx.snap_gen_scheduler,
         ) {
             error!(
                 "schedule snapshot failed";
@@ -4758,7 +4758,7 @@ pub struct Builder<EK: KvEngine> {
     cfg: Arc<VersionTrack<Config>>,
     coprocessor_host: CoprocessorHost<EK>,
     importer: Arc<SstImporter<EK>>,
-    region_scheduler: Scheduler<RegionTask<<EK as KvEngine>::Snapshot>>,
+    snap_gen_scheduler: Scheduler<SnapGenTask<EK::Snapshot>>,
     engine: EK,
     sender: Box<dyn Notifier<EK>>,
     router: ApplyRouter<EK>,
@@ -4777,7 +4777,7 @@ impl<EK: KvEngine> Builder<EK> {
             cfg: builder.cfg.clone(),
             coprocessor_host: builder.coprocessor_host.clone(),
             importer: builder.importer.clone(),
-            region_scheduler: builder.region_scheduler.clone(),
+            snap_gen_scheduler: builder.snap_gen_scheduler.clone(),
             engine: builder.engines.kv.clone(),
             sender,
             router,
@@ -4801,7 +4801,7 @@ where
                 self.tag.clone(),
                 self.coprocessor_host.clone(),
                 self.importer.clone(),
-                self.region_scheduler.clone(),
+                self.snap_gen_scheduler.clone(),
                 self.engine.clone(),
                 self.router.clone(),
                 self.sender.clone_box(),
@@ -4827,7 +4827,7 @@ where
             cfg: self.cfg.clone(),
             coprocessor_host: self.coprocessor_host.clone(),
             importer: self.importer.clone(),
-            region_scheduler: self.region_scheduler.clone(),
+            snap_gen_scheduler: self.snap_gen_scheduler.clone(),
             engine: self.engine.clone(),
             sender: self.sender.clone_box(),
             router: self.router.clone(),
@@ -5146,7 +5146,7 @@ mod tests {
             msg::WriteResponse,
             peer_storage::RAFT_INIT_LOG_INDEX,
             simple_write::{SimpleWriteEncoder, SimpleWriteReqEncoder},
-            Config, RegionTask,
+            Config, SnapGenTask,
         },
     };
 
@@ -5469,7 +5469,7 @@ mod tests {
         let sender = Box::new(TestNotifier { tx });
         let (_tmp, engine) = create_tmp_engine("apply-basic");
         let (_dir, importer) = create_tmp_importer("apply-basic");
-        let (region_scheduler, mut snapshot_rx) = dummy_scheduler();
+        let (snap_gen_scheduler, mut snapshot_rx) = dummy_scheduler();
         let cfg = Arc::new(VersionTrack::new(Config::default()));
         let (router, mut system) = create_apply_batch_system(&cfg.value(), None);
         let pending_create_peers = Arc::new(Mutex::new(HashMap::default()));
@@ -5478,7 +5478,7 @@ mod tests {
             cfg,
             coprocessor_host: CoprocessorHost::<KvTestEngine>::default(),
             importer,
-            region_scheduler,
+            snap_gen_scheduler,
             sender,
             engine,
             router: router.clone(),
@@ -5571,7 +5571,7 @@ mod tests {
         };
         let apply_state_key = keys::apply_state_key(2);
         let apply_state = match snapshot_rx.recv_timeout(Duration::from_secs(3)) {
-            Ok(Some(RegionTask::Gen { kv_snap, .. })) => kv_snap
+            Ok(Some(SnapGenTask::Gen { kv_snap, .. })) => kv_snap
                 .get_msg_cf(CF_RAFT, &apply_state_key)
                 .unwrap()
                 .unwrap(),
@@ -6038,7 +6038,7 @@ mod tests {
             .register_query_observer(1, BoxQueryObserver::new(obs.clone()));
 
         let (tx, rx) = mpsc::channel();
-        let (region_scheduler, _) = dummy_scheduler();
+        let (snap_gen_scheduler, _) = dummy_scheduler();
         let sender = Box::new(TestNotifier { tx });
         let cfg = Arc::new(VersionTrack::new(Config::default()));
         let (router, mut system) = create_apply_batch_system(&cfg.value(), None);
@@ -6047,7 +6047,7 @@ mod tests {
             tag: "test-store".to_owned(),
             cfg,
             sender,
-            region_scheduler,
+            snap_gen_scheduler,
             coprocessor_host: host,
             importer: importer.clone(),
             engine: engine.clone(),
@@ -6377,7 +6377,7 @@ mod tests {
             .register_query_observer(1, BoxQueryObserver::new(obs.clone()));
 
         let (tx, rx) = mpsc::channel();
-        let (region_scheduler, _) = dummy_scheduler();
+        let (snap_gen_scheduler, _) = dummy_scheduler();
         let sender = Box::new(TestNotifier { tx });
         let mut config = Config::default();
         config.enable_v2_compatible_learner = true;
@@ -6388,7 +6388,7 @@ mod tests {
             tag: "test-store".to_owned(),
             cfg,
             sender,
-            region_scheduler,
+            snap_gen_scheduler,
             coprocessor_host: host,
             importer: importer.clone(),
             engine: engine.clone(),
@@ -6722,7 +6722,7 @@ mod tests {
             .register_query_observer(1, BoxQueryObserver::new(obs));
 
         let (tx, rx) = mpsc::channel();
-        let (region_scheduler, _) = dummy_scheduler();
+        let (snap_gen_scheduler, _) = dummy_scheduler();
         let sender = Box::new(TestNotifier { tx });
         let cfg = Arc::new(VersionTrack::new(Config::default()));
         let (router, mut system) = create_apply_batch_system(&cfg.value(), None);
@@ -6731,7 +6731,7 @@ mod tests {
             tag: "test-store".to_owned(),
             cfg: cfg.clone(),
             sender,
-            region_scheduler,
+            snap_gen_scheduler,
             coprocessor_host: host,
             importer,
             engine,
@@ -6808,7 +6808,7 @@ mod tests {
             .register_query_observer(1, BoxQueryObserver::new(obs));
 
         let (tx, rx) = mpsc::channel();
-        let (region_scheduler, _) = dummy_scheduler();
+        let (snap_gen_scheduler, _) = dummy_scheduler();
         let sender = Box::new(TestNotifier { tx });
         let cfg = {
             let mut cfg = Config::default();
@@ -6822,7 +6822,7 @@ mod tests {
             tag: "test-store".to_owned(),
             cfg,
             sender,
-            region_scheduler,
+            snap_gen_scheduler,
             coprocessor_host: host,
             importer: importer.clone(),
             engine: engine.clone(),
@@ -6991,7 +6991,7 @@ mod tests {
             .register_query_observer(1, BoxQueryObserver::new(obs));
 
         let (tx, rx) = mpsc::channel();
-        let (region_scheduler, _) = dummy_scheduler();
+        let (snap_gen_scheduler, _) = dummy_scheduler();
         let sender = Box::new(TestNotifier { tx });
         let cfg = {
             let mut cfg = Config::default();
@@ -7005,7 +7005,7 @@ mod tests {
             tag: "test-store".to_owned(),
             cfg,
             sender,
-            region_scheduler,
+            snap_gen_scheduler,
             coprocessor_host: host,
             importer,
             engine,
@@ -7089,7 +7089,7 @@ mod tests {
             .register_query_observer(1, BoxQueryObserver::new(obs.clone()));
 
         let (tx, rx) = mpsc::channel();
-        let (region_scheduler, _) = dummy_scheduler();
+        let (snap_gen_scheduler, _) = dummy_scheduler();
         let sender = Box::new(TestNotifier { tx });
         let cfg = Config::default();
         let (router, mut system) = create_apply_batch_system(&cfg, None);
@@ -7098,7 +7098,7 @@ mod tests {
             tag: "test-exec-observer".to_owned(),
             cfg: Arc::new(VersionTrack::new(cfg)),
             sender,
-            region_scheduler,
+            snap_gen_scheduler,
             coprocessor_host: host,
             importer: importer.clone(),
             engine: engine.clone(),
@@ -7314,7 +7314,7 @@ mod tests {
             .register_cmd_observer(1, BoxCmdObserver::new(obs));
 
         let (tx, rx) = mpsc::channel();
-        let (region_scheduler, _) = dummy_scheduler();
+        let (snap_gen_scheduler, _) = dummy_scheduler();
         let sender = Box::new(TestNotifier { tx });
         let cfg = Config::default();
         let (router, mut system) = create_apply_batch_system(&cfg, None);
@@ -7323,7 +7323,7 @@ mod tests {
             tag: "test-store".to_owned(),
             cfg: Arc::new(VersionTrack::new(cfg)),
             sender,
-            region_scheduler,
+            snap_gen_scheduler,
             coprocessor_host: host,
             importer,
             engine,
@@ -7595,7 +7595,7 @@ mod tests {
         obs.cmd_sink = Some(Arc::new(Mutex::new(sink)));
         host.registry
             .register_cmd_observer(1, BoxCmdObserver::new(obs));
-        let (region_scheduler, _) = dummy_scheduler();
+        let (snap_gen_scheduler, _) = dummy_scheduler();
         let cfg = Arc::new(VersionTrack::new(Config::default()));
         let (router, mut system) = create_apply_batch_system(&cfg.value(), None);
         let pending_create_peers = Arc::new(Mutex::new(HashMap::default()));
@@ -7604,7 +7604,7 @@ mod tests {
             cfg,
             sender,
             importer,
-            region_scheduler,
+            snap_gen_scheduler,
             coprocessor_host: host,
             engine: engine.clone(),
             router: router.clone(),
@@ -7815,7 +7815,7 @@ mod tests {
         let (tx, apply_res_rx) = mpsc::channel();
         let sender = Box::new(TestNotifier { tx });
         let coprocessor_host = CoprocessorHost::<KvTestEngine>::default();
-        let (region_scheduler, _) = dummy_scheduler();
+        let (snap_gen_scheduler, _) = dummy_scheduler();
         let cfg = Arc::new(VersionTrack::new(Config::default()));
         let (router, mut system) = create_apply_batch_system(&cfg.value(), None);
         let pending_create_peers = Arc::new(Mutex::new(HashMap::default()));
@@ -7824,7 +7824,7 @@ mod tests {
             cfg,
             sender,
             importer,
-            region_scheduler,
+            snap_gen_scheduler,
             coprocessor_host,
             engine: engine.clone(),
             router: router.clone(),
@@ -7939,7 +7939,7 @@ mod tests {
             .register_query_observer(1, BoxQueryObserver::new(ApplyObserver::default()));
 
         let (tx, rx) = mpsc::channel();
-        let (region_scheduler, _) = dummy_scheduler();
+        let (snap_gen_scheduler, _) = dummy_scheduler();
         let sender = Box::new(TestNotifier { tx });
         let cfg = Arc::new(VersionTrack::new(Config::default()));
         let (router, mut system) = create_apply_batch_system(&cfg.value(), None);
@@ -7948,7 +7948,7 @@ mod tests {
             tag: "flashback_need_to_be_applied".to_owned(),
             cfg,
             sender,
-            region_scheduler,
+            snap_gen_scheduler,
             coprocessor_host: host,
             importer,
             engine: engine.clone(),
